Revert "fix: move shared and keys"
