+++ backend/.gitignore
@@ -2,4 +2,4 @@
 node_modules/
 
 # env vars
-#.env
+.env
