--- backend/lesson-service/.env
@@ -1,12 +1,0 @@
-# Server configuration
-PORT=3001
-
-# Firebase configuration
-FIREBASE_API_KEY=AIzaSyAtxjsRtpuIbcEg48AaZfsh-bOuC_2ydTk
-FIREBASE_AUTH_DOMAIN=codelingo-1.firebaseapp.com
-FIREBASE_DATABASE_URL=https://codelingo-1-default-rtdb.firebaseio.com
-FIREBASE_PROJECT_ID=codelingo-1
-FIREBASE_STORAGE_BUCKET=codelingo-1.firebasestorage.app
-FIREBASE_MESSAGING_SENDER_ID=391880921910
-FIREBASE_APP_ID=1:391880921910:web:189a794e40b50cd2c05997
-GEMINI_API_KEY="AIzaSyAfUSzeCsPFuH8cEiGbD-kxpaqMRSjLGNs"
