--- backend/lesson-service/keys/serviceAccountKey.json
@@ -1,14 +1,0 @@
-{
-    "type": "service_account",
-    "project_id": "codelingo-1",
-    "private_key_id": "abff2caff9e79b4689ce33ea5440cfceadf1c648",
-    "private_key": "-----BEGIN PRIVATE KEY-----\nMIIEvgIBADANBgkqhkiG9w0BAQEFAASCBKgwggSkAgEAAoIBAQDTq0o9QQPheLSS\nF3BivrbBiZunxK6hSNybCnzH/nJxKwN/jsPiTxlvmY8VqjeuFCS3QUW3guI7Sdxy\nPbxnB7cICG165OFHJFJCJ1RhCIfPbMkt7u8XYB8bwoE1+CXH22QZe/4CBWvvNrXL\nmlieyIeda1asqKSJlJjcBQB5UyZj/Ylx8rgGyrx7JY6yg13g5egi0J7xx2TRXCbV\nIasaRxUNTH/rr7fqpnlRkqBBMQKyKHMmZDVJQ0KYFC9nOuNnroWzzjWpsDtzIOOv\nlhclKZ1sVj3fvHAEQJ9FkuqkihExa2KQyW5iiQFResgo5IqFW2r4jq/MZ6vuvSqF\nJ8C2wxNnAgMBAAECggEAIEWVZBFtjgrhPQTl1ahOJOPeH44VMfHx8JGAxN0+hZKR\nTfluVCrLLgCzmBIDuz7G+kmxX3FNHMz+F0ziVLMQbOEcEAilcDtRb1SfwMuUKUPk\nyGQNf/rfw1O8b4YN/JKuiFQnPhelhStKPGij6lGYm5+p71km/PLy+1in4deTmtre\nBHoBldBea40jwY25N7kBVaThTw/WGR8KE2WRR1tTDygfKP2hQSr54EBkEFFkn6xi\nZOIgAVho8Qjgc+yTp7IMnLSvcEihCryG2inSyCtJR8jJEWn3vB0VZ0EQUOz8CiSn\ng1909sst2BSIUK+WmftsTphtRPJNQNV/Vjfe5cF3wQKBgQDsE2mHwkBGpdmTb/Ir\nJANm9CoGkq0lbsWh46LCSVU2Zvp/coa1M2UD7yQkxunlCdxolQRyy71QMro03vnK\nxVPt2tby0vly/wyiwjRdtgihy1I7eHDId8eDEEPFTLaZ0WE3Tqs2R1Zbqs68iytT\nWTw3ytbqdqU2tUEy96jdIqUz0wKBgQDliI2K3c3I8rhs1/KwprEUyngYX3utoZXS\nj/Eo2AAnwh0qDW5vlR+Pvf4NMFnYjqaVF04B8LPM2bl0KGhBUARl3kL91/Ep7hhP\n7FbwhLMVySK7bftHeENe6lTlV+UrxAtbOBnoApk3F4vbHGTJ4uXalRp25q4cMJmu\nsf/inVSpnQKBgQDAhZw8UJgmhT6I1HSD5hy4K0rSjW+S5WFbAs9K7o8FjLmgkf91\nanJo4O9y5pz8YixTRaOZErpS2uusDmNbj81AXoWJBbDUrVbarNIT0k371Fo9kZ04\nfyY/plX+Yu9PRceoWQmbfFXa79MlQ8GTeWR4g3z9PMDDetGvBPW7UXO0bQKBgBfn\nTZ49lRWtOUmUf6Cn618X6sPXBcFCR6EC/oFbZUC+tVzQw0R5iizlA1UKSQEFchGq\nuKP5FjBqKCcRojHTDHIzYFShLWtJcVy5ZyBrIV9OcACG5pk9y/5Rdv/fVG1nYCti\nv0frJVuYZyctYM2phHhr+LKTOWSUJx0LSnOmIM/9AoGBAIE+jaof9Opwq9jSKIe3\nXEtsq/Y4Hyz+BKuthsy4Y9JB2oxgJUWXKWKux+d7lPP7cZ4URJ7vDO73/8F2EEUd\ngRwlvRTFEBNJvXEqjPDRKN4zSHCIrxCeBsvlJzMjFEkGsDYRjUZUad5BL2cCuycu\nsoq5odPpRH4SWcx6kQUBhhmK\n-----END PRIVATE KEY-----\n",
-    "client_email": "firebase-adminsdk-fbsvc@codelingo-1.iam.gserviceaccount.com",
-    "client_id": "110009014991450474670",
-    "auth_uri": "https://accounts.google.com/o/oauth2/auth",
-    "token_uri": "https://oauth2.googleapis.com/token",
-    "auth_provider_x509_cert_url": "https://www.googleapis.com/oauth2/v1/certs",
-    "client_x509_cert_url": "https://www.googleapis.com/robot/v1/metadata/x509/firebase-adminsdk-fbsvc%40codelingo-1.iam.gserviceaccount.com",
-    "universe_domain": "googleapis.com"
-  }
-
